refactor the code to separate the output bindings and utility routines.
