+++ NwNsgProject/host.json
@@ -1,3 +1,3 @@
 {
-  "functions": [ "Stage1BlobTrigger", "Stage2QueueTrigger", "Stage3QueueTrigger" ]
+  "functions": [ "Stage3QueueTrigger" ]
 }
